+++ src/components/orders-payments/index.ts
@@ -8,3 +8,4 @@ export { default as ProductName } from './ProductName';
 export { default as ProductAmoutInfo } from './ProductAmoutInfo';
 export { default as ProductOptionInfo } from './ProductOptionInfo';
 export { default as OrderTableInfoSection } from './OrderTableInfoSection';
+export { default as OrderStatusBadge } from './OrderStatusBadge';
