+++ src/types/badgeType.ts
@@ -1,0 +1,21 @@
+export type BadgeVariant =
+  | 'secondary'
+  | 'secondaryOutline'
+  | 'secondaryOutlineSquare'
+  | 'secondaryDot'
+  | 'primary'
+  | 'primaryOutline'
+  | 'primaryOutlineSquare'
+  | 'primaryDot'
+  | 'success'
+  | 'successOutline'
+  | 'successOutlineSquare'
+  | 'successDot'
+  | 'warning'
+  | 'warningOutline'
+  | 'warningOutlineSquare'
+  | 'warningDot'
+  | 'danger'
+  | 'dangerOutline'
+  | 'dangerOutlineSquare'
+  | 'dangerDot';
